Ensure license is accepted on remote bootstrap target
When performing the initial `chef-client` converge
on a bootstrapped host, ensure that we set the license
environment variable that indicates the license was accepted.

This will allow the run to continue under Chef 15, which requires
license acceptance.

I used the environment variable instead of the CLI flag
because older versions of Chef do not support the CLI flag.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>